refactor: redefine enroll course webhook
